+++ frontend/components/AccordionContainer/AccordionSection.jsx
@@ -1,11 +1,15 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import ArrowIcon from '@shopgate/pwa-ui-shared/icons/ArrowIcon';
 import RatingStars from '@shopgate/pwa-ui-shared/RatingStars';
+import AccordionIcon from './AccordionIcon';
 import styles from './style';
 
+import getConfig from '../../helpers/getConfig';
+
+const { showLastSectionBottomBorder } = getConfig();
+
 /**
- * A simple AccoridonSection Component
+ * A simple AccordionSection Component
  */
 class AccordionSection extends Component {
   static propTypes = {
@@ -13,6 +17,7 @@ class AccordionSection extends Component {
     label: PropTypes.string.isRequired,
     onClick: PropTypes.func.isRequired,
     children: PropTypes.instanceOf(Object),
+    isLast: PropTypes.bool,
     isRating: PropTypes.bool,
     rating: PropTypes.instanceOf(Object),
   }
@@ -21,6 +26,7 @@ class AccordionSection extends Component {
     children: null,
     isRating: false,
     rating: null,
+    isLast: false,
   };
 
   onClick = () => {
@@ -35,12 +41,12 @@ class AccordionSection extends Component {
     const {
       onClick,
       props: {
-        isOpen, label, isRating, rating,
+        isOpen, label, isRating, rating, isLast,
       },
     } = this;
 
     return (
-      <div>
+      <div className={showLastSectionBottomBorder && isLast ? styles.sectionBorder : null}>
         <button onClick={onClick} className={styles.sectionButton} type="button">
           <div className={styles.sectionTitleWrapper}>
             <div className={styles.sectionTitle}>
@@ -54,10 +60,7 @@ class AccordionSection extends Component {
                 value={rating && rating.average ? rating.average : 0}
               />
             }
-            <div className={styles.arrow}>
-              {!isOpen && <span><ArrowIcon className={styles.arrowDown} /></span>}
-              {isOpen && <span><ArrowIcon className={styles.arrowUp} /></span>}
-            </div>
+            <AccordionIcon isOpen={isOpen} />
           </div>
         </button>
         {isOpen && (